Fix hang during "Database Installation in Progress..."
The application hung while "Database Installation in Progress..." was
displayed in the progress panel.  The database was not installed.

A '404 Not Found' error was caused by the inclusion of
'/wp-admin/install.php' in the URL created in the 'install_wp'  and
'install_theme' functions.   The incorrect URL was of the form:

'http://mysite/wp-quick-install//wp-admin/install.php?action=install_wp'

The correct URL is:

'http://mysite/wp-quick-install?action=install_wp'